+++ lib/src/domain_certificate_file_path.dart
@@ -1,15 +1,17 @@
 import 'dart:io';
 
-import '../shelf_letsencrypt.dart';
 import 'certs_handler.dart';
 
 /// A [DomainCertificate] implementation using file paths.
 class DomainCertificateFilePath extends DomainCertificate {
-  DomainCertificateFilePath(List<String> super.domains, this.fullChainFilePath,
-      this.privateKeyFilePath);
+  /// The fullChain file path.
   final String fullChainFilePath;
 
+  /// The private key file path.
   final String privateKeyFilePath;
+
+  DomainCertificateFilePath(List<String> super.domains, this.fullChainFilePath,
+      this.privateKeyFilePath);
 
   @override
   String get fullChainPEM => File(fullChainFilePath).readAsStringSync();