+++ test/shelf_letsencrypt_test.dart
@@ -25,7 +25,7 @@ void main() {
       expect(await certificatesHandler.getAccountPEMKeyPair(), isNull);
       expect(
           await certificatesHandler.getDomainPEMKeyPair(domain.name), isNull);
-      expect(await certificatesHandler.buildSecurityContext([domain]), isNull);
+      expect(await certificatesHandler.buildSecurityContexts([domain]), isNull);
 
       final accountPEMKeyPair =
           await certificatesHandler.ensureAccountPEMKeyPair();
@@ -176,7 +176,7 @@ void main() {
       expect(error, isNotNull);
       expect(
           error?.message,
-          allOf(contains('No previous SecureContext'),
+          allOf(contains("Can't load all `SecurityContext`s"),
               contains("can't request")));
     });